+++ src/components/Ledger/LedgerList/LedgerListFilterConfig.js
@@ -1,0 +1,15 @@
+import { LEDGER_FILTERS } from '../constants';
+
+// eslint-disable-next-line import/prefer-default-export
+export const filterConfig = [
+  {
+    name: LEDGER_FILTERS.STATUS,
+    cql: LEDGER_FILTERS.STATUS,
+    values: [],
+  },
+  {
+    name: LEDGER_FILTERS.ACQUISITIONS_UNIT,
+    cql: LEDGER_FILTERS.ACQUISITIONS_UNIT,
+    values: [],
+  },
+];
