+++ src/components/Ledger/LedgerList/LedgerListFilters/index.js
@@ -1,0 +1,1 @@
+export { default } from './LedgerListFiltersContainer';
